pyproject.toml
[project]
name = "propertymanagement"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "flask>=3.1.1",
    "psycopg2-binary>=2.9.10",
]
